improve downscaling for single frame images
The default downscaling provided by the GPU can result in noisy
artifacts on highly detailed images.

This commit employs a cubic Catmull-Rom sampling filter for the
case where we have a single frame image that is being reduced
in size.  This isn't the highest quality filter but strikes
a good balance with speed vs appearance and is strictly better
than the GPU texture sampling options that I could try.