term: revise image geometry calculation
The texture coordinates into the source image weren't quite right;
previously we'd do this math:

num_cells = (image_size / cell_size).ceil()
delta = num_cells / image_size

that would result in the image being stretched to fit across the rounded
up number of cells, leading to a distored image.

This commit changes the delta calculation to be based on the remaining
number of pixels in a given dimension relative to the cell size.

refs: #1300
refs: #1270